rfp-accelerator-customer-model/requirements.txt
gunicorn==21.2.0
google-cloud-discoveryengine==0.12.1
langchain==0.3.0
langchain-community==0.3.0
langchain-google-vertexai==2.0.1
html2text==2020.1.16
google-cloud-aiplatform==1.67.1
fastapi==0.110.0
typing-extensions==4.8.0
uvicorn==0.25.0
nest-asyncio==1.5.1
google-cloud-logging==3.9.0
google-cloud-translate==3.13.0
google-cloud-secret-manager==2.17.0
google-api-python-client==2.146.0
google-cloud-bigquery==3.25.0
docarray==0.40.0
scikit-learn==1.3.2
pydantic==2.9.2
SQLAlchemy==2.0.7
cloud-sql-python-connector==1.5.0
sqlalchemy-pytds==0.3.5
pymysql==1.1.0
pandas==2.0.3
tqdm==4.66.1
openpyxl==3.1.2
pyarrow==14.0.0
tabulate==0.9.0
PyJWT==2.8.0
Markdown==3.5.2
beautifulsoup4==4.12.3
PyPDF2==3.0.1
python-multipart==0.0.9
pdfminer.six==20231228
docx2txt==0.8